moved non-generic routes to example
