+++ examples/express.ts
@@ -14,6 +14,7 @@ import InMemoryStorage from './implementation/InMemoryStorage';
 import JsonStorage from './implementation/JsonStorage';
 import User from './implementation/User';
 import startPageRenderer from './startPageRenderer';
+import expressRoutes from './expressRoutes';
 
 const start = async () => {
     const h5pEditor = new H5P.H5PEditor(
@@ -54,6 +55,8 @@ const start = async () => {
             path.resolve('h5p/editor')
         )
     );
+
+    server.use('/h5p', expressRoutes(h5pEditor));
 
     server.get('/', startPageRenderer(h5pEditor));
